refactor: ChessBoard의 Rank 수정
