+++ src/test/java/domain/RankTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import chess.domain.Rank;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
